Optimisation of thread implementation in crash area
